test: split tests to fils
