+++ test/lib/index.spec.js
@@ -234,2085 +234,6 @@ describe('The API', function() {
 });
 
 describe('Mocking the file system', function() {
-  if (fs.access && fs.accessSync && process.getuid && process.getgid) {
-    // TODO: drop condition when dropping Node < 0.12 support
-    // TODO: figure out how fs.access() works on Windows (without gid/uid)
-
-    describe('fs.access(path[, mode], callback)', function() {
-      beforeEach(function() {
-        mock({
-          'path/to/accessible/file': 'can access',
-          'path/to/000': mock.file({
-            mode: parseInt('0000', 8),
-            content: 'no permissions'
-          }),
-          'path/to/111': mock.file({
-            mode: parseInt('0111', 8),
-            content: 'execute only'
-          }),
-          'path/to/write/only': mock.file({
-            mode: parseInt('0222', 8),
-            content: 'write only'
-          }),
-          'path/to/333': mock.file({
-            mode: parseInt('0333', 8),
-            content: 'write and execute'
-          }),
-          'path/to/444': mock.file({
-            mode: parseInt('0444', 8),
-            content: 'read only'
-          }),
-          'path/to/555': mock.file({
-            mode: parseInt('0555', 8),
-            content: 'read and execute'
-          }),
-          'path/to/666': mock.file({
-            mode: parseInt('0666', 8),
-            content: 'read and write'
-          }),
-          'path/to/777': mock.file({
-            mode: parseInt('0777', 8),
-            content: 'read, write, and execute'
-          }),
-          unreadable: mock.directory({
-            mode: parseInt('0000', 8),
-            items: {
-              'readable-child': mock.file({
-                mode: parseInt('0777', 8),
-                content: 'read, write, and execute'
-              })
-            }
-          })
-        });
-      });
-      afterEach(mock.restore);
-
-      it('works for an accessible file', function(done) {
-        fs.access('path/to/accessible/file', done);
-      });
-
-      withPromise.it('promise works for an accessible file', function(done) {
-        fs.promises.access('path/to/accessible/file').then(done, done);
-      });
-
-      it('works 000 (and no mode arg)', function(done) {
-        fs.access('path/to/000', done);
-      });
-
-      withPromise.it('promise works 000 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/000').then(done, done);
-      });
-
-      it('works F_OK and 000', function(done) {
-        fs.access('path/to/000', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 000', function(done) {
-        fs.promises.access('path/to/000', fs.F_OK).then(done, done);
-      });
-
-      it('generates EACCES for R_OK and 000', function(done) {
-        fs.access('path/to/000', fs.R_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for R_OK and 000', function(
-        done
-      ) {
-        fs.promises.access('path/to/000', fs.R_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('generates EACCES for W_OK and 000', function(done) {
-        fs.access('path/to/000', fs.W_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for W_OK and 000', function(
-        done
-      ) {
-        fs.promises.access('path/to/000', fs.W_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('generates EACCES for X_OK and 000', function(done) {
-        fs.access('path/to/000', fs.X_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for X_OK and 000', function(
-        done
-      ) {
-        fs.promises.access('path/to/000', fs.X_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('works 111 (and no mode arg)', function(done) {
-        fs.access('path/to/111', done);
-      });
-
-      withPromise.it('promise works 111 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/111').then(done, done);
-      });
-
-      it('works F_OK and 111', function(done) {
-        fs.access('path/to/111', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 111', function(done) {
-        fs.promises.access('path/to/111', fs.F_OK).then(done, done);
-      });
-
-      it('works X_OK and 111', function(done) {
-        fs.access('path/to/111', fs.X_OK, done);
-      });
-
-      withPromise.it('promise works X_OK and 111', function(done) {
-        fs.promises.access('path/to/111', fs.X_OK).then(done, done);
-      });
-
-      it('generates EACCES for R_OK and 111', function(done) {
-        fs.access('path/to/111', fs.R_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for R_OK and 111', function(
-        done
-      ) {
-        fs.promises.access('path/to/111', fs.R_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('generates EACCES for W_OK and 111', function(done) {
-        fs.access('path/to/111', fs.W_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for W_OK and 111', function(
-        done
-      ) {
-        fs.promises.access('path/to/111', fs.W_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('works for 222 (and no mode arg)', function(done) {
-        fs.access('path/to/write/only', done);
-      });
-
-      withPromise.it('promise works for 222 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/write/only').then(done, done);
-      });
-
-      it('works F_OK and 222', function(done) {
-        fs.access('path/to/write/only', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 222', function(done) {
-        fs.promises.access('path/to/write/only', fs.F_OK).then(done, done);
-      });
-
-      it('works W_OK and 222', function(done) {
-        fs.access('path/to/write/only', fs.W_OK, done);
-      });
-
-      withPromise.it('promise works W_OK and 222', function(done) {
-        fs.promises.access('path/to/write/only', fs.W_OK).then(done, done);
-      });
-
-      it('generates EACCES for R_OK and 222', function(done) {
-        fs.access('path/to/write/only', fs.R_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for R_OK and 222', function(
-        done
-      ) {
-        fs.promises.access('path/to/write/only', fs.R_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('generates EACCES for X_OK and 222', function(done) {
-        fs.access('path/to/write/only', fs.X_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for X_OK and 222', function(
-        done
-      ) {
-        fs.promises.access('path/to/write/only', fs.X_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('works for 333 (and no mode arg)', function(done) {
-        fs.access('path/to/333', done);
-      });
-
-      withPromise.it('promise works for 333 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/333').then(done, done);
-      });
-
-      it('works F_OK and 333', function(done) {
-        fs.access('path/to/333', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 333', function(done) {
-        fs.promises.access('path/to/333', fs.F_OK).then(done, done);
-      });
-
-      it('works W_OK and 333', function(done) {
-        fs.access('path/to/333', fs.W_OK, done);
-      });
-
-      withPromise.it('promise works W_OK and 333', function(done) {
-        fs.promises.access('path/to/333', fs.W_OK).then(done, done);
-      });
-
-      it('works X_OK and 333', function(done) {
-        fs.access('path/to/333', fs.X_OK, done);
-      });
-
-      withPromise.it('promise works X_OK and 333', function(done) {
-        fs.promises.access('path/to/333', fs.X_OK).then(done, done);
-      });
-
-      it('works X_OK | W_OK and 333', function(done) {
-        fs.access('path/to/333', fs.X_OK | fs.W_OK, done);
-      });
-
-      withPromise.it('promise works X_OK | W_OK and 333', function(done) {
-        fs.promises.access('path/to/333', fs.X_OK | fs.W_OK).then(done, done);
-      });
-
-      it('generates EACCES for R_OK and 333', function(done) {
-        fs.access('path/to/333', fs.R_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for R_OK and 333', function(
-        done
-      ) {
-        fs.promises.access('path/to/333', fs.R_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('works for 444 (and no mode arg)', function(done) {
-        fs.access('path/to/444', done);
-      });
-
-      withPromise.it('promise works for 444 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/444').then(done, done);
-      });
-
-      it('works F_OK and 444', function(done) {
-        fs.access('path/to/444', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 444', function(done) {
-        fs.promises.access('path/to/444', fs.F_OK).then(done, done);
-      });
-
-      it('works R_OK and 444', function(done) {
-        fs.access('path/to/444', fs.R_OK, done);
-      });
-
-      withPromise.it('promise works R_OK and 444', function(done) {
-        fs.promises.access('path/to/444', fs.R_OK).then(done, done);
-      });
-
-      it('generates EACCES for W_OK and 444', function(done) {
-        fs.access('path/to/444', fs.W_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for W_OK and 444', function(
-        done
-      ) {
-        fs.promises.access('path/to/444', fs.W_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('generates EACCES for X_OK and 444', function(done) {
-        fs.access('path/to/444', fs.X_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for X_OK and 444', function(
-        done
-      ) {
-        fs.promises.access('path/to/444', fs.X_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('works for 555 (and no mode arg)', function(done) {
-        fs.access('path/to/555', done);
-      });
-
-      withPromise.it('promise works for 555 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/555').then(done, done);
-      });
-
-      it('works F_OK and 555', function(done) {
-        fs.access('path/to/555', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 555', function(done) {
-        fs.promises.access('path/to/555', fs.F_OK).then(done, done);
-      });
-
-      it('works R_OK and 555', function(done) {
-        fs.access('path/to/555', fs.R_OK, done);
-      });
-
-      withPromise.it('promise works R_OK and 555', function(done) {
-        fs.promises.access('path/to/555', fs.R_OK).then(done, done);
-      });
-
-      it('works X_OK and 555', function(done) {
-        fs.access('path/to/555', fs.X_OK, done);
-      });
-
-      withPromise.it('promise works X_OK and 555', function(done) {
-        fs.promises.access('path/to/555', fs.X_OK).then(done, done);
-      });
-
-      it('works R_OK | X_OK and 555', function(done) {
-        fs.access('path/to/555', fs.R_OK | fs.X_OK, done);
-      });
-
-      withPromise.it('promise works R_OK | X_OK and 555', function(done) {
-        fs.promises.access('path/to/555', fs.R_OK | fs.X_OK).then(done, done);
-      });
-
-      it('generates EACCES for W_OK and 555', function(done) {
-        fs.access('path/to/555', fs.W_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for W_OK and 555', function(
-        done
-      ) {
-        fs.promises.access('path/to/555', fs.W_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('works for 666 (and no mode arg)', function(done) {
-        fs.access('path/to/666', done);
-      });
-
-      withPromise.it('promise works for 666 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/666').then(done, done);
-      });
-
-      it('works F_OK and 666', function(done) {
-        fs.access('path/to/666', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 666', function(done) {
-        fs.promises.access('path/to/666', fs.F_OK).then(done, done);
-      });
-
-      it('works R_OK and 666', function(done) {
-        fs.access('path/to/666', fs.R_OK, done);
-      });
-
-      withPromise.it('promise works R_OK and 666', function(done) {
-        fs.promises.access('path/to/666', fs.R_OK).then(done, done);
-      });
-
-      it('works W_OK and 666', function(done) {
-        fs.access('path/to/666', fs.W_OK, done);
-      });
-
-      withPromise.it('promise works W_OK and 666', function(done) {
-        fs.promises.access('path/to/666', fs.W_OK).then(done, done);
-      });
-
-      it('works R_OK | W_OK and 666', function(done) {
-        fs.access('path/to/666', fs.R_OK | fs.W_OK, done);
-      });
-
-      withPromise.it('promise works R_OK | W_OK and 666', function(done) {
-        fs.promises.access('path/to/666', fs.R_OK | fs.W_OK).then(done, done);
-      });
-
-      it('generates EACCES for X_OK and 666', function(done) {
-        fs.access('path/to/666', fs.X_OK, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it('promise generates EACCES for X_OK and 666', function(
-        done
-      ) {
-        fs.promises.access('path/to/666', fs.X_OK).then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EACCES');
-            done();
-          }
-        );
-      });
-
-      it('works for 777 (and no mode arg)', function(done) {
-        fs.access('path/to/777', done);
-      });
-
-      withPromise.it('promise works for 777 (and no mode arg)', function(done) {
-        fs.promises.access('path/to/777').then(done, done);
-      });
-
-      it('works F_OK and 777', function(done) {
-        fs.access('path/to/777', fs.F_OK, done);
-      });
-
-      withPromise.it('promise works F_OK and 777', function(done) {
-        fs.promises.access('path/to/777', fs.F_OK).then(done, done);
-      });
-
-      it('works R_OK and 777', function(done) {
-        fs.access('path/to/777', fs.R_OK, done);
-      });
-
-      withPromise.it('promise works R_OK and 777', function(done) {
-        fs.promises.access('path/to/777', fs.R_OK).then(done, done);
-      });
-
-      it('works W_OK and 777', function(done) {
-        fs.access('path/to/777', fs.W_OK, done);
-      });
-
-      withPromise.it('promise works W_OK and 777', function(done) {
-        fs.promises.access('path/to/777', fs.W_OK).then(done, done);
-      });
-
-      it('works X_OK and 777', function(done) {
-        fs.access('path/to/777', fs.X_OK, done);
-      });
-
-      withPromise.it('promise works X_OK and 777', function(done) {
-        fs.promises.access('path/to/777', fs.X_OK).then(done, done);
-      });
-
-      it('works X_OK | W_OK and 777', function(done) {
-        fs.access('path/to/777', fs.X_OK | fs.W_OK, done);
-      });
-
-      withPromise.it('promise works X_OK | W_OK and 777', function(done) {
-        fs.promises.access('path/to/777', fs.X_OK | fs.W_OK).then(done, done);
-      });
-
-      it('works X_OK | R_OK and 777', function(done) {
-        fs.access('path/to/777', fs.X_OK | fs.R_OK, done);
-      });
-
-      withPromise.it('promise works X_OK | R_OK and 777', function(done) {
-        fs.promises.access('path/to/777', fs.X_OK | fs.R_OK).then(done, done);
-      });
-
-      it('works R_OK | W_OK and 777', function(done) {
-        fs.access('path/to/777', fs.R_OK | fs.W_OK, done);
-      });
-
-      withPromise.it('promise works R_OK | W_OK and 777', function(done) {
-        fs.promises.access('path/to/777', fs.R_OK | fs.W_OK).then(done, done);
-      });
-
-      it('works R_OK | W_OK | X_OK and 777', function(done) {
-        fs.access('path/to/777', fs.R_OK | fs.W_OK | fs.X_OK, done);
-      });
-
-      withPromise.it('promise works R_OK | W_OK | X_OK and 777', function(
-        done
-      ) {
-        fs.promises
-          .access('path/to/777', fs.R_OK | fs.W_OK | fs.X_OK)
-          .then(done, done);
-      });
-
-      it('generates EACCESS for F_OK and an unreadable parent', function(done) {
-        fs.access('unreadable/readable-child', function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EACCES');
-          done();
-        });
-      });
-
-      withPromise.it(
-        'promise generates EACCESS for F_OK and an unreadable parent',
-        function(done) {
-          fs.promises.access('unreadable/readable-child').then(
-            function() {
-              assert.fail('should not succeed.');
-              done();
-            },
-            function(err) {
-              assert.instanceOf(err, Error);
-              assert.equal(err.code, 'EACCES');
-              done();
-            }
-          );
-        }
-      );
-    });
-
-    describe('fs.accessSync(path[, mode])', function() {
-      beforeEach(function() {
-        mock({
-          'path/to/777': mock.file({
-            mode: parseInt('0777', 8),
-            content: 'all access'
-          }),
-          'path/to/000': mock.file({
-            mode: parseInt('0000', 8),
-            content: 'no permissions'
-          }),
-          'broken-link': mock.symlink({path: './path/to/nothing'}),
-          'circular-link': mock.symlink({path: './loop-link'}),
-          'loop-link': mock.symlink({path: './circular-link'})
-        });
-      });
-      afterEach(mock.restore);
-
-      it('works for an accessible file', function() {
-        fs.accessSync('path/to/777');
-        fs.accessSync('path/to/777', fs.F_OK);
-        fs.accessSync('path/to/777', fs.X_OK);
-        fs.accessSync('path/to/777', fs.W_OK);
-        fs.accessSync('path/to/777', fs.X_OK | fs.W_OK);
-        fs.accessSync('path/to/777', fs.R_OK);
-        fs.accessSync('path/to/777', fs.X_OK | fs.R_OK);
-        fs.accessSync('path/to/777', fs.W_OK | fs.R_OK);
-        fs.accessSync('path/to/777', fs.X_OK | fs.W_OK | fs.R_OK);
-      });
-
-      it('throws EACCESS for broken link', function() {
-        assert.throws(function() {
-          fs.accessSync('broken-link');
-        });
-      });
-
-      it('throws ELOOP for circular link', function() {
-        assert.throws(function() {
-          fs.accessSync('circular-link');
-        });
-      });
-
-      it('throws EACCESS for all but F_OK for 000', function() {
-        fs.accessSync('path/to/000');
-        assert.throws(function() {
-          fs.accessSync('path/to/000', fs.X_OK);
-        });
-        assert.throws(function() {
-          fs.accessSync('path/to/000', fs.W_OK);
-        });
-        assert.throws(function() {
-          fs.accessSync('path/to/000', fs.X_OK | fs.W_OK);
-        });
-        assert.throws(function() {
-          fs.accessSync('path/to/000', fs.R_OK);
-        });
-        assert.throws(function() {
-          fs.accessSync('path/to/000', fs.X_OK | fs.R_OK);
-        });
-        assert.throws(function() {
-          fs.accessSync('path/to/000', fs.W_OK | fs.R_OK);
-        });
-        assert.throws(function() {
-          fs.accessSync('path/to/000', fs.X_OK | fs.W_OK | fs.R_OK);
-        });
-      });
-    });
-  }
-
-  if (fs.copyFile && fs.copyFileSync) {
-    describe('fs.copyFile(src, dest[, flags], callback)', function() {
-      beforeEach(function() {
-        mock({
-          'path/to/src.txt': 'file content',
-          'path/to/other.txt': 'other file content',
-          empty: {}
-        });
-      });
-      afterEach(mock.restore);
-
-      it('copies a file to an empty directory', function(done) {
-        fs.copyFile('path/to/src.txt', 'empty/dest.txt', function(err) {
-          assert.isTrue(!err);
-          assert.isTrue(fs.existsSync('empty/dest.txt'));
-          assert.equal(
-            String(fs.readFileSync('empty/dest.txt')),
-            'file content'
-          );
-          done();
-        });
-      });
-
-      withPromise.it('promise copies a file to an empty directory', function(
-        done
-      ) {
-        fs.promises
-          .copyFile('path/to/src.txt', 'empty/dest.txt')
-          .then(function() {
-            assert.isTrue(fs.existsSync('empty/dest.txt'));
-            assert.equal(
-              String(fs.readFileSync('empty/dest.txt')),
-              'file content'
-            );
-            done();
-          }, done);
-      });
-
-      it('truncates dest file if it exists', function(done) {
-        fs.copyFile('path/to/src.txt', 'path/to/other.txt', function(err) {
-          assert.isTrue(!err);
-          assert.equal(
-            String(fs.readFileSync('path/to/other.txt')),
-            'file content'
-          );
-          done();
-        });
-      });
-
-      withPromise.it('promise truncates dest file if it exists', function(
-        done
-      ) {
-        fs.promises
-          .copyFile('path/to/src.txt', 'path/to/other.txt')
-          .then(function() {
-            assert.equal(
-              String(fs.readFileSync('path/to/other.txt')),
-              'file content'
-            );
-            done();
-          }, done);
-      });
-
-      it('throws if dest exists and exclusive', function(done) {
-        fs.copyFile(
-          'path/to/src.txt',
-          'path/to/other.txt',
-          fs.constants.COPYFILE_EXCL,
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EEXIST');
-            done();
-          }
-        );
-      });
-
-      withPromise.it('promise throws if dest exists and exclusive', function(
-        done
-      ) {
-        fs.promises
-          .copyFile(
-            'path/to/src.txt',
-            'path/to/other.txt',
-            fs.constants.COPYFILE_EXCL
-          )
-          .then(
-            function() {
-              assert.fail('should not succeed.');
-              done();
-            },
-            function(err) {
-              assert.instanceOf(err, Error);
-              assert.equal(err.code, 'EEXIST');
-              done();
-            }
-          );
-      });
-
-      it('fails if src does not exist', function(done) {
-        fs.copyFile('path/to/bogus.txt', 'empty/dest.txt', function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'ENOENT');
-          done();
-        });
-      });
-
-      withPromise.it('promise fails if src does not exist', function(done) {
-        fs.promises.copyFile('path/to/bogus.txt', 'empty/dest.txt').then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'ENOENT');
-            done();
-          }
-        );
-      });
-
-      it('fails if dest path does not exist', function(done) {
-        fs.copyFile('path/to/src.txt', 'path/nope/dest.txt', function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'ENOENT');
-          done();
-        });
-      });
-
-      withPromise.it('promise fails if dest path does not exist', function(
-        done
-      ) {
-        fs.promises.copyFile('path/to/src.txt', 'path/nope/dest.txt').then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'ENOENT');
-            done();
-          }
-        );
-      });
-
-      it('fails if dest is a directory', function(done) {
-        fs.copyFile('path/to/src.txt', 'empty', function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EISDIR');
-          done();
-        });
-      });
-
-      withPromise.it('promise fails if dest is a directory', function(done) {
-        fs.promises.copyFile('path/to/src.txt', 'empty').then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EISDIR');
-            done();
-          }
-        );
-      });
-    });
-  }
-
-  describe('fs.rename(oldPath, newPath, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/a.bin': bufferFrom([1, 2, 3]),
-        empty: {},
-        nested: {
-          dir: mock.directory({
-            mtime: new Date(1),
-            items: {'file.txt': ''}
-          })
-        }
-      });
-    });
-    afterEach(mock.restore);
-
-    it('allows files to be renamed', function(done) {
-      fs.rename('path/to/a.bin', 'path/to/b.bin', function(err) {
-        assert.isTrue(!err);
-        assert.isFalse(fs.existsSync('path/to/a.bin'));
-        assert.isTrue(fs.existsSync('path/to/b.bin'));
-        done();
-      });
-    });
-
-    withPromise.it('promise allows files to be renamed', function(done) {
-      fs.promises.rename('path/to/a.bin', 'path/to/b.bin').then(function() {
-        assert.isFalse(fs.existsSync('path/to/a.bin'));
-        assert.isTrue(fs.existsSync('path/to/b.bin'));
-        done();
-      }, done);
-    });
-
-    it('updates mtime of parent directory', function(done) {
-      const oldTime = fs.statSync('nested/dir').mtime;
-      fs.rename('nested/dir/file.txt', 'nested/dir/renamed.txt', function(err) {
-        assert.isTrue(!err);
-        assert.isFalse(fs.existsSync('nested/dir/file.txt'));
-        assert.isTrue(fs.existsSync('nested/dir/renamed.txt'));
-        const newTime = fs.statSync('nested/dir').mtime;
-        assert.isTrue(newTime > oldTime);
-        done();
-      });
-    });
-
-    withPromise.it('promise updates mtime of parent directory', function(done) {
-      const oldTime = fs.statSync('nested/dir').mtime;
-      fs.promises
-        .rename('nested/dir/file.txt', 'nested/dir/renamed.txt')
-        .then(function() {
-          assert.isFalse(fs.existsSync('nested/dir/file.txt'));
-          assert.isTrue(fs.existsSync('nested/dir/renamed.txt'));
-          const newTime = fs.statSync('nested/dir').mtime;
-          assert.isTrue(newTime > oldTime);
-          done();
-        }, done);
-    });
-
-    it('calls callback with error if old path does not exist', function(done) {
-      fs.rename('bogus', 'empty', function(err) {
-        assert.instanceOf(err, Error);
-        done();
-      });
-    });
-
-    withPromise.it(
-      'promise calls callback with error if old path does not exist',
-      function(done) {
-        fs.promises.rename('bogus', 'empty').then(
-          function() {
-            assert.fail('Should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            done();
-          }
-        );
-      }
-    );
-
-    it('overwrites existing files', function(done) {
-      fs.rename('path/to/a.bin', 'nested/dir/file.txt', function(err) {
-        assert.isTrue(!err);
-        assert.isFalse(fs.existsSync('path/to/a.bin'));
-        assert.isTrue(fs.existsSync('nested/dir/file.txt'));
-        done();
-      });
-    });
-
-    withPromise.it('promise overwrites existing files', function(done) {
-      fs.promises
-        .rename('path/to/a.bin', 'nested/dir/file.txt')
-        .then(function() {
-          assert.isFalse(fs.existsSync('path/to/a.bin'));
-          assert.isTrue(fs.existsSync('nested/dir/file.txt'));
-          done();
-        }, done);
-    });
-
-    it('allows directories to be renamed', function(done) {
-      fs.rename('path/to', 'path/foo', function(err) {
-        assert.isTrue(!err);
-        assert.isFalse(fs.existsSync('path/to'));
-        assert.isTrue(fs.existsSync('path/foo'));
-        assert.deepEqual(fs.readdirSync('path/foo'), ['a.bin']);
-        done();
-      });
-    });
-
-    withPromise.it('promise allows directories to be renamed', function(done) {
-      fs.promises.rename('path/to', 'path/foo').then(function() {
-        assert.isFalse(fs.existsSync('path/to'));
-        assert.isTrue(fs.existsSync('path/foo'));
-        assert.deepEqual(fs.readdirSync('path/foo'), ['a.bin']);
-        done();
-      }, done);
-    });
-
-    it('calls callback with error if new directory not empty', function(done) {
-      fs.rename('path', 'nested', function(err) {
-        assert.instanceOf(err, Error);
-        done();
-      });
-    });
-
-    withPromise.it(
-      'promise calls callback with error if new directory not empty',
-      function(done) {
-        fs.promises.rename('path', 'nested').then(
-          function() {
-            assert.fail('Should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            done();
-          }
-        );
-      }
-    );
-  });
-
-  describe('fs.renameSync(oldPath, newPath)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/a.bin': bufferFrom([1, 2, 3]),
-        empty: {},
-        nested: {
-          dir: {
-            'file.txt': ''
-          }
-        },
-        link: mock.symlink({path: './path/to/a.bin'})
-      });
-    });
-    afterEach(mock.restore);
-
-    it('allows files to be renamed', function() {
-      fs.renameSync('path/to/a.bin', 'path/to/b.bin');
-      assert.isFalse(fs.existsSync('path/to/a.bin'));
-      assert.isTrue(fs.existsSync('path/to/b.bin'));
-    });
-
-    it('overwrites existing files', function() {
-      fs.renameSync('path/to/a.bin', 'nested/dir/file.txt');
-      assert.isFalse(fs.existsSync('path/to/a.bin'));
-      assert.isTrue(fs.existsSync('nested/dir/file.txt'));
-    });
-
-    it('allows directories to be renamed', function() {
-      fs.renameSync('path/to', 'path/foo');
-      assert.isFalse(fs.existsSync('path/to'));
-      assert.isTrue(fs.existsSync('path/foo'));
-      assert.deepEqual(fs.readdirSync('path/foo'), ['a.bin']);
-    });
-
-    it('replaces existing directories (if empty)', function() {
-      fs.renameSync('path/to', 'empty');
-      assert.isFalse(fs.existsSync('path/to'));
-      assert.isTrue(fs.existsSync('empty'));
-      assert.deepEqual(fs.readdirSync('empty'), ['a.bin']);
-    });
-
-    it('renames symbolic links', function() {
-      fs.renameSync('link', 'renamed');
-      assert.isTrue(fs.existsSync('renamed'));
-      assert.isFalse(fs.existsSync('link'));
-      assert.isTrue(fs.existsSync('path/to/a.bin'));
-    });
-
-    it('throws if old path does not exist', function() {
-      assert.throws(function() {
-        fs.renameSync('bogus', 'empty');
-      });
-    });
-
-    it('throws if new path basename is not directory', function() {
-      assert.throws(function() {
-        fs.renameSync('path/to/a.bin', 'bogus/a.bin');
-      });
-    });
-
-    it('throws if new dir is not empty dir', function() {
-      assert.throws(function() {
-        fs.renameSync('path/to', 'nested');
-      });
-    });
-  });
-
-  describe('fs.stat(path, callback)', function() {
-    beforeEach(function() {
-      mock({
-        '/path/to/file.txt': mock.file({
-          ctime: new Date(1),
-          mtime: new Date(2),
-          atime: new Date(3),
-          uid: 42,
-          gid: 43
-        }),
-        '/dir/symlink': mock.symlink({path: '/path/to/file.txt'}),
-        '/empty': {}
-      });
-    });
-    afterEach(mock.restore);
-
-    xit('creates an instance of fs.Stats', function(done) {
-      fs.stat('/path/to/file.txt', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.instanceOf(stats, fs.Stats);
-        done();
-      });
-    });
-
-    withPromise.xit('promise creates an instance of fs.Stats', function(done) {
-      fs.promises.stat('/path/to/file.txt').then(function(stats) {
-        assert.instanceOf(stats, fs.Stats);
-        done();
-      }, done);
-    });
-
-    it('identifies files', function(done) {
-      fs.stat('/path/to/file.txt', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isTrue(stats.isFile());
-        assert.isFalse(stats.isDirectory());
-        done();
-      });
-    });
-
-    withPromise.it('promise identifies files', function(done) {
-      fs.promises.stat('/path/to/file.txt').then(function(stats) {
-        assert.isTrue(stats.isFile());
-        assert.isFalse(stats.isDirectory());
-        done();
-      }, done);
-    });
-
-    it('identifies directories', function(done) {
-      fs.stat('/empty', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isTrue(stats.isDirectory());
-        assert.isFalse(stats.isFile());
-        done();
-      });
-    });
-
-    withPromise.it('promise identifies directories', function(done) {
-      fs.promises.stat('/empty').then(function(stats) {
-        assert.isTrue(stats.isDirectory());
-        assert.isFalse(stats.isFile());
-        done();
-      }, done);
-    });
-
-    it('provides file stats', function(done) {
-      fs.stat('/path/to/file.txt', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.equal(stats.ctime.getTime(), 1);
-        assert.equal(stats.mtime.getTime(), 2);
-        assert.equal(stats.atime.getTime(), 3);
-        assert.equal(stats.uid, 42);
-        assert.equal(stats.gid, 43);
-        assert.equal(stats.nlink, 1);
-        assert.isNumber(stats.rdev);
-        done();
-      });
-    });
-
-    withPromise.it('promise provides file stats', function(done) {
-      fs.promises.stat('/path/to/file.txt').then(function(stats) {
-        assert.equal(stats.ctime.getTime(), 1);
-        assert.equal(stats.mtime.getTime(), 2);
-        assert.equal(stats.atime.getTime(), 3);
-        assert.equal(stats.uid, 42);
-        assert.equal(stats.gid, 43);
-        assert.equal(stats.nlink, 1);
-        assert.isNumber(stats.rdev);
-        done();
-      }, done);
-    });
-
-    it('includes blocks and blksize in stats', function(done) {
-      fs.stat('/path/to/file.txt', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isNumber(stats.blocks);
-        assert.isNumber(stats.blksize);
-        done();
-      });
-    });
-
-    withPromise.it('promise includes blocks and blksize in stats', function(
-      done
-    ) {
-      fs.promises.stat('/path/to/file.txt').then(function(stats) {
-        assert.isNumber(stats.blocks);
-        assert.isNumber(stats.blksize);
-        done();
-      }, done);
-    });
-
-    it('provides directory stats', function(done) {
-      fs.stat('/path', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.instanceOf(stats.ctime, Date);
-        assert.instanceOf(stats.mtime, Date);
-        assert.instanceOf(stats.atime, Date);
-        if (process.getuid) {
-          assert.isNumber(stats.uid);
-        } else {
-          assert.isNaN(stats.uid);
-        }
-        if (process.getgid) {
-          assert.isNumber(stats.gid);
-        } else {
-          assert.isNaN(stats.gid);
-        }
-        assert.equal(stats.nlink, 3);
-        assert.isNumber(stats.rdev);
-        done();
-      });
-    });
-
-    withPromise.it('promise provides directory stats', function(done) {
-      fs.promises.stat('/path').then(function(stats) {
-        assert.instanceOf(stats.ctime, Date);
-        assert.instanceOf(stats.mtime, Date);
-        assert.instanceOf(stats.atime, Date);
-        if (process.getuid) {
-          assert.isNumber(stats.uid);
-        } else {
-          assert.isNaN(stats.uid);
-        }
-        if (process.getgid) {
-          assert.isNumber(stats.gid);
-        } else {
-          assert.isNaN(stats.gid);
-        }
-        assert.equal(stats.nlink, 3);
-        assert.isNumber(stats.rdev);
-        done();
-      }, done);
-    });
-
-    it('includes blocks and blksize in directory stats', function(done) {
-      fs.stat('/path', function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isNumber(stats.blocks);
-        assert.isNumber(stats.blksize);
-        done();
-      });
-    });
-
-    withPromise.it(
-      'promise includes blocks and blksize in directory stats',
-      function(done) {
-        fs.promises.stat('/path').then(function(stats) {
-          assert.isNumber(stats.blocks);
-          assert.isNumber(stats.blksize);
-          done();
-        }, done);
-      }
-    );
-  });
-
-  describe('fs.fstat(fd, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content',
-        empty: {}
-      });
-    });
-    afterEach(mock.restore);
-
-    it('accepts a file descriptor for a file (r)', function(done) {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      fs.fstat(fd, function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isTrue(stats.isFile());
-        assert.equal(stats.size, 12);
-        done();
-      });
-    });
-
-    withPromise.it('promise accepts a file descriptor for a file (r)', function(
-      done
-    ) {
-      fs.promises
-        .open('path/to/file.txt', 'r')
-        .then(function(fd) {
-          return fd.stat();
-        })
-        .then(function(stats) {
-          assert.isTrue(stats.isFile());
-          assert.equal(stats.size, 12);
-          done();
-        }, done);
-    });
-
-    it('accepts a file descriptor for a directory (r)', function(done) {
-      const fd = fs.openSync('path/to', 'r');
-      fs.fstat(fd, function(err, stats) {
-        if (err) {
-          return done(err);
-        }
-        assert.isTrue(stats.isDirectory());
-        assert.isTrue(stats.size > 0);
-        done();
-      });
-    });
-
-    withPromise.it(
-      'promise accepts a file descriptor for a directory (r)',
-      function(done) {
-        fs.promises
-          .open('path/to', 'r')
-          .then(function(fd) {
-            return fd.stat();
-          })
-          .then(function(stats) {
-            assert.isTrue(stats.isDirectory());
-            assert.isTrue(stats.size > 0);
-            done();
-          }, done);
-      }
-    );
-
-    it('fails for bad file descriptor', function(done) {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      fs.closeSync(fd);
-      fs.fstat(fd, function(err, stats) {
-        assert.instanceOf(err, Error);
-        assert.equal(err.code, 'EBADF');
-        done();
-      });
-    });
-
-    withPromise.it('promise fails for bad file descriptor', function(done) {
-      fs.promises
-        .open('path/to/file.txt', 'r')
-        .then(function(fd) {
-          return fd.close().then(function() {
-            return fd.stat();
-          });
-        })
-        .then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EBADF');
-            done();
-          }
-        );
-    });
-  });
-
-  describe('fs.fstatSync(fd)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content',
-        empty: {}
-      });
-    });
-    afterEach(mock.restore);
-
-    it('accepts a file descriptor for a file (r)', function() {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      const stats = fs.fstatSync(fd);
-      assert.isTrue(stats.isFile());
-      assert.equal(stats.size, 12);
-    });
-
-    it('accepts a file descriptor for a directory (r)', function() {
-      const fd = fs.openSync('path/to', 'r');
-      const stats = fs.fstatSync(fd);
-      assert.isTrue(stats.isDirectory());
-      assert.isTrue(stats.size > 0);
-    });
-
-    it('fails for bad file descriptor', function() {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      fs.closeSync(fd);
-      assert.throws(function() {
-        fs.fstatSync(fd);
-      });
-    });
-  });
-
-  describe('fs.exists(path, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/a.bin': bufferFrom([1, 2, 3]),
-        empty: {},
-        nested: {
-          dir: {
-            'file.txt': ''
-          }
-        }
-      });
-    });
-    afterEach(mock.restore);
-
-    it('calls with true if file exists', function(done) {
-      fs.exists(path.join('path', 'to', 'a.bin'), function(exists) {
-        assert.isTrue(exists);
-        done();
-      });
-    });
-
-    it('calls with true if directory exists', function(done) {
-      fs.exists('path', function(exists) {
-        assert.isTrue(exists);
-        done();
-      });
-    });
-
-    it('calls with true if empty directory exists', function(done) {
-      fs.exists('empty', function(exists) {
-        assert.isTrue(exists);
-        done();
-      });
-    });
-
-    it('calls with true if nested directory exists', function(done) {
-      fs.exists(path.join('nested', 'dir'), function(exists) {
-        assert.isTrue(exists);
-        done();
-      });
-    });
-
-    it('calls with true if file exists', function(done) {
-      fs.exists(path.join('path', 'to', 'a.bin'), function(exists) {
-        assert.isTrue(exists);
-        done();
-      });
-    });
-
-    it('calls with true if empty file exists', function(done) {
-      fs.exists(path.join('nested', 'dir', 'file.txt'), function(exists) {
-        assert.isTrue(exists);
-        done();
-      });
-    });
-
-    it('calls with false for bogus path', function(done) {
-      fs.exists(path.join('bogus', 'path'), function(exists) {
-        assert.isFalse(exists);
-        done();
-      });
-    });
-
-    it('calls with false for bogus path (II)', function(done) {
-      fs.exists(path.join('nested', 'dir', 'none'), function(exists) {
-        assert.isFalse(exists);
-        done();
-      });
-    });
-  });
-
-  describe('fs.existsSync(path)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/a.bin': bufferFrom([1, 2, 3]),
-        empty: {},
-        nested: {
-          dir: {
-            'file.txt': ''
-          }
-        }
-      });
-    });
-    afterEach(mock.restore);
-
-    it('returns true if file exists', function() {
-      assert.isTrue(fs.existsSync(path.join('path', 'to', 'a.bin')));
-    });
-
-    it('returns true if directory exists', function() {
-      assert.isTrue(fs.existsSync('path'));
-    });
-
-    it('returns true if empty directory exists', function() {
-      assert.isTrue(fs.existsSync('empty'));
-    });
-
-    it('returns true if nested directory exists', function() {
-      assert.isTrue(fs.existsSync(path.join('nested', 'dir')));
-    });
-
-    it('returns true if file exists', function() {
-      assert.isTrue(fs.existsSync(path.join('path', 'to', 'a.bin')));
-    });
-
-    it('returns true if empty file exists', function() {
-      assert.isTrue(fs.existsSync(path.join('nested', 'dir', 'file.txt')));
-    });
-
-    it('returns false for bogus path', function() {
-      assert.isFalse(fs.existsSync(path.join('bogus', 'path')));
-    });
-
-    it('returns false for bogus path (II)', function() {
-      assert.isFalse(fs.existsSync(path.join('nested', 'dir', 'none')));
-    });
-  });
-
-  describe('fs.readdir(path, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content',
-        nested: {
-          sub: {
-            dir: {
-              'one.txt': 'one content',
-              'two.txt': 'two content',
-              empty: {}
-            }
-          }
-        }
-      });
-    });
-    afterEach(mock.restore);
-
-    it('lists directory contents', function(done) {
-      fs.readdir(path.join('path', 'to'), function(err, items) {
-        assert.isNull(err);
-        assert.isArray(items);
-        assert.deepEqual(items, ['file.txt']);
-        done();
-      });
-    });
-
-    withPromise.it('promise lists directory contents', function(done) {
-      fs.promises.readdir(path.join('path', 'to')).then(function(items) {
-        assert.isArray(items);
-        assert.deepEqual(items, ['file.txt']);
-        done();
-      }, done);
-    });
-
-    it('lists nested directory contents', function(done) {
-      fs.readdir(path.join('nested', 'sub', 'dir'), function(err, items) {
-        assert.isNull(err);
-        assert.isArray(items);
-        assert.deepEqual(items, ['empty', 'one.txt', 'two.txt']);
-        done();
-      });
-    });
-
-    withPromise.it('promise lists nested directory contents', function(done) {
-      fs.promises
-        .readdir(path.join('nested', 'sub', 'dir'))
-        .then(function(items) {
-          assert.isArray(items);
-          assert.deepEqual(items, ['empty', 'one.txt', 'two.txt']);
-          done();
-        }, done);
-    });
-
-    it('calls with an error for bogus path', function(done) {
-      fs.readdir('bogus', function(err, items) {
-        assert.instanceOf(err, Error);
-        assert.isUndefined(items);
-        done();
-      });
-    });
-
-    withPromise.it('promise calls with an error for bogus path', function(
-      done
-    ) {
-      fs.promises.readdir('bogus').then(
-        function() {
-          assert.fail('should not succeed.');
-          done();
-        },
-        function(err) {
-          assert.instanceOf(err, Error);
-          done();
-        }
-      );
-    });
-  });
-
-  describe('fs.readdirSync(path)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content',
-        nested: {
-          sub: {
-            dir: {
-              'one.txt': 'one content',
-              'two.txt': 'two content',
-              empty: {}
-            }
-          }
-        }
-      });
-    });
-    afterEach(mock.restore);
-
-    it('lists directory contents', function() {
-      const items = fs.readdirSync(path.join('path', 'to'));
-      assert.isArray(items);
-      assert.deepEqual(items, ['file.txt']);
-    });
-
-    it('lists nested directory contents', function() {
-      const items = fs.readdirSync(path.join('nested', 'sub', 'dir'));
-      assert.isArray(items);
-      assert.deepEqual(items, ['empty', 'one.txt', 'two.txt']);
-    });
-
-    it('throws for bogus path', function() {
-      assert.throws(function() {
-        fs.readdirSync('bogus');
-      });
-    });
-  });
-
-  describe('fs.open(path, flags, [mode], callback)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content',
-        nested: {
-          sub: {
-            dir: {
-              'one.txt': 'one content',
-              'two.txt': 'two content',
-              empty: {}
-            }
-          }
-        }
-      });
-    });
-    afterEach(mock.restore);
-
-    it('opens an existing file for reading (r)', function(done) {
-      fs.open('nested/sub/dir/one.txt', 'r', function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        assert.isNumber(fd);
-        done();
-      });
-    });
-
-    withPromise.it('promise opens an existing file for reading (r)', function(
-      done
-    ) {
-      fs.promises.open('nested/sub/dir/one.txt', 'r').then(function(fd) {
-        assert.isNumber(fd.fd);
-        done();
-      }, done);
-    });
-
-    it('fails if file does not exist (r)', function(done) {
-      fs.open('bogus.txt', 'r', function(err, fd) {
-        assert.instanceOf(err, Error);
-        assert.equal(err.code, 'ENOENT');
-        done();
-      });
-    });
-
-    withPromise.it('promise fails if file does not exist (r)', function(done) {
-      fs.promises.open('bogus.txt', 'r').then(
-        function() {
-          assert.fail('should not succeed.');
-          done();
-        },
-        function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'ENOENT');
-          done();
-        }
-      );
-    });
-
-    it('creates a new file for writing (w)', function(done) {
-      fs.open('path/to/new.txt', 'w', parseInt('0666', 8), function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        assert.isNumber(fd);
-        assert.isTrue(fs.existsSync('path/to/new.txt'));
-        done();
-      });
-    });
-
-    withPromise.it('promise creates a new file for writing (w)', function(
-      done
-    ) {
-      fs.promises
-        .open('path/to/new.txt', 'w', parseInt('0666', 8))
-        .then(function(fd) {
-          assert.isNumber(fd.fd);
-          assert.isTrue(fs.existsSync('path/to/new.txt'));
-          done();
-        }, done);
-    });
-
-    it('opens an existing file for writing (w)', function(done) {
-      fs.open('path/to/file.txt', 'w', parseInt('0666', 8), function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        assert.isNumber(fd);
-        done();
-      });
-    });
-
-    withPromise.it('promise opens an existing file for writing (w)', function(
-      done
-    ) {
-      fs.promises
-        .open('path/to/file.txt', 'w', parseInt('0666', 8))
-        .then(function(fd) {
-          assert.isNumber(fd.fd);
-          done();
-        }, done);
-    });
-
-    it('fails if file exists (wx)', function(done) {
-      fs.open('path/to/file.txt', 'wx', parseInt('0666', 8), function(err, fd) {
-        assert.instanceOf(err, Error);
-        assert.equal(err.code, 'EEXIST');
-        done();
-      });
-    });
-
-    withPromise.it('promise fails if file exists (wx)', function(done) {
-      fs.promises.open('path/to/file.txt', 'wx', parseInt('0666', 8)).then(
-        function() {
-          assert.fail('should not succeed.');
-          done();
-        },
-        function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EEXIST');
-          done();
-        }
-      );
-    });
-  });
-
-  describe('fs.openSync(path, flags, [mode])', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content',
-        nested: {
-          sub: {
-            dir: {
-              'one.txt': 'one content',
-              'two.txt': 'two content',
-              empty: {}
-            }
-          }
-        }
-      });
-    });
-    afterEach(mock.restore);
-
-    it('opens an existing file for reading (r)', function() {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      assert.isNumber(fd);
-    });
-
-    it('fails if file does not exist (r)', function() {
-      assert.throws(function() {
-        fs.openSync('bogus.txt', 'r');
-      });
-    });
-
-    it('creates a new file for writing (w)', function() {
-      const fd = fs.openSync('nested/sub/new.txt', 'w', parseInt('0666', 8));
-      assert.isNumber(fd);
-      assert.isTrue(fs.existsSync('nested/sub/new.txt'));
-    });
-
-    it('opens an existing file for writing (w)', function() {
-      const fd = fs.openSync('path/to/one.txt', 'w', parseInt('0666', 8));
-      assert.isNumber(fd);
-    });
-
-    it('fails if file exists (wx)', function() {
-      assert.throws(function() {
-        fs.openSync('path/to/file.txt', 'wx', parseInt('0666', 8));
-      });
-    });
-  });
-
-  describe('fs.close(fd, callback)', function() {
-    beforeEach(function() {
-      mock({dir: {}});
-    });
-    afterEach(mock.restore);
-
-    it('closes a file descriptor', function(done) {
-      const fd = fs.openSync('dir/file.txt', 'w');
-      fs.close(fd, function(err) {
-        done(err);
-      });
-    });
-
-    withPromise.it('promise closes a file descriptor', function(done) {
-      fs.promises
-        .open('dir/file.txt', 'w')
-        .then(function(fd) {
-          return fd.close();
-        })
-        .then(done, done);
-    });
-
-    it('fails for closed file descriptors', function(done) {
-      const fd = fs.openSync('dir/file.txt', 'w');
-      fs.close(fd, function(err) {
-        if (err) {
-          return done(err);
-        }
-        fs.close(fd, function(err) {
-          assert.instanceOf(err, Error);
-          assert.equal(err.code, 'EBADF');
-          done();
-        });
-      });
-    });
-
-    withPromise.it('promise fails for closed file descriptors', function(done) {
-      fs.promises
-        .open('dir/file.txt', 'w')
-        .then(function(fd) {
-          return fd.close().then(function() {
-            return fd.close();
-          });
-        })
-        .then(
-          function() {
-            assert.fail('should not succeed.');
-            done();
-          },
-          function(err) {
-            assert.instanceOf(err, Error);
-            assert.equal(err.code, 'EBADF');
-            done();
-          }
-        );
-    });
-  });
-
-  describe('fs.closeSync(fd)', function() {
-    beforeEach(function() {
-      mock({dir: {}});
-    });
-    afterEach(mock.restore);
-
-    it('closes a file descriptor', function() {
-      const fd = fs.openSync('dir/file.txt', 'w');
-      fs.closeSync(fd);
-    });
-
-    it('fails for closed file descriptors', function() {
-      const fd = fs.openSync('dir/file.txt', 'w');
-      fs.closeSync(fd);
-      assert.throws(function() {
-        fs.closeSync(fd);
-      });
-    });
-  });
-
-  describe('fs.read(fd, buffer, offset, length, position, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content'
-      });
-    });
-    afterEach(mock.restore);
-
-    it('allows file contents to be read', function(done) {
-      fs.open('path/to/file.txt', 'r', function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        const buffer = bufferAlloc(12);
-        fs.read(fd, buffer, 0, 12, 0, function(err2, bytesRead, buf) {
-          if (err2) {
-            return done(err2);
-          }
-          assert.equal(bytesRead, 12);
-          assert.equal(buf, buffer);
-          assert.equal(String(buffer), 'file content');
-          done();
-        });
-      });
-    });
-
-    it('allows file contents to be read w/ offset', function(done) {
-      fs.open('path/to/file.txt', 'r', function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        const buffer = bufferAlloc(12);
-        fs.read(fd, buffer, 5, 7, 0, function(err2, bytesRead, buf) {
-          if (err2) {
-            return done(err2);
-          }
-          assert.equal(bytesRead, 7);
-          assert.equal(buf, buffer);
-          assert.equal(String(buffer.slice(5)), 'file co');
-          done();
-        });
-      });
-    });
-
-    it('allows file contents to be read w/ length', function(done) {
-      fs.open('path/to/file.txt', 'r', function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        const buffer = bufferAlloc(12);
-        fs.read(fd, buffer, 0, 4, 0, function(err2, bytesRead, buf) {
-          if (err2) {
-            return done(err2);
-          }
-          assert.equal(bytesRead, 4);
-          assert.equal(buf, buffer);
-          assert.equal(String(buffer.slice(0, 4)), 'file');
-          done();
-        });
-      });
-    });
-
-    it('allows file contents to be read w/ offset & length', function(done) {
-      fs.open('path/to/file.txt', 'r', function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        const buffer = bufferAlloc(12);
-        fs.read(fd, buffer, 2, 4, 0, function(err2, bytesRead, buf) {
-          if (err2) {
-            return done(err2);
-          }
-          assert.equal(bytesRead, 4);
-          assert.equal(buf, buffer);
-          assert.equal(String(buffer.slice(2, 6)), 'file');
-          done();
-        });
-      });
-    });
-
-    it('allows file contents to be read w/ position', function(done) {
-      fs.open('path/to/file.txt', 'r', function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        const buffer = bufferAlloc(7);
-        fs.read(fd, buffer, 0, 7, 5, function(err2, bytesRead, buf) {
-          if (err2) {
-            return done(err2);
-          }
-          assert.equal(bytesRead, 7);
-          assert.equal(buf, buffer);
-          assert.equal(String(buffer), 'content');
-          done();
-        });
-      });
-    });
-
-    it('allows read w/ offset, length, & position', function(done) {
-      fs.open('path/to/file.txt', 'r', function(err, fd) {
-        if (err) {
-          return done(err);
-        }
-        const buffer = bufferAlloc(12);
-        fs.read(fd, buffer, 2, 7, 5, function(err2, bytesRead, buf) {
-          if (err2) {
-            return done(err2);
-          }
-          assert.equal(bytesRead, 7);
-          assert.equal(buf, buffer);
-          assert.equal(String(buffer.slice(2, 9)), 'content');
-          done();
-        });
-      });
-    });
-
-    it('fails for closed file descriptor', function(done) {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      fs.closeSync(fd);
-      fs.read(fd, bufferAlloc(12), 0, 12, 0, function(err, bytesRead, buf) {
-        assert.instanceOf(err, Error);
-        assert.equal(0, bytesRead);
-        done();
-      });
-    });
-
-    it('fails if not open for reading', function(done) {
-      const fd = fs.openSync('path/to/file.txt', 'w');
-      fs.read(fd, bufferAlloc(12), 0, 12, 0, function(err, bytesRead, buf) {
-        assert.instanceOf(err, Error);
-        assert.equal(0, bytesRead);
-        done();
-      });
-    });
-  });
-
-  describe('fs.readSync(fd, buffer, offset, length, position)', function() {
-    beforeEach(function() {
-      mock({
-        'path/to/file.txt': 'file content'
-      });
-    });
-    afterEach(mock.restore);
-
-    it('allows a file to be read synchronously', function() {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      const buffer = bufferAlloc(12);
-      const read = fs.readSync(fd, buffer, 0, 12, 0);
-      assert.equal(read, 12);
-      assert.equal(String(buffer), 'file content');
-    });
-
-    it('allows a file to be read in two parts', function() {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      const first = bufferAlloc(4);
-      fs.readSync(fd, first, 0, 4, 0);
-      assert.equal(String(first), 'file');
-
-      const second = bufferAlloc(7);
-      fs.readSync(fd, second, 0, 7, 5);
-      assert.equal(String(second), 'content');
-    });
-
-    it('treats null position as current position', function() {
-      const fd = fs.openSync('path/to/file.txt', 'r');
-      const first = bufferAlloc(4);
-      fs.readSync(fd, first, 0, 4, null);
-      assert.equal(String(first), 'file');
-
-      // consume the space
-      assert.equal(fs.readSync(fd, bufferAlloc(1), 0, 1, null), 1);
-
-      const second = bufferAlloc(7);
-      fs.readSync(fd, second, 0, 7, null);
-      assert.equal(String(second), 'content');
-    });
-  });
 
   describe('fs.readFile(filename, [options], callback)', function() {
     // this is provided by fs.open, fs.fstat, and fs.read
